Исправление округления

+++ src/main/java/laba4/OSAGOwithTrailer.java
@@ -9,7 +9,7 @@ public class OSAGOwithTrailer extends OSAGO {
 		setConf(category_try, city_try, cbm_try, age_try, experience_try, cm_try);
 		promo = 1;
 		int tbInt = (int) (((btMax - btMin)*100*staf+btMin)*ct*cbm*csv*cm*cpr*promo);
-		float tb = tbInt/100;
+		float tb = ((float)tbInt)/100;
 		return tb;
 	}
